+++ NEWS.md
@@ -2,6 +2,13 @@
 # easy-samba news
 Stay updated on `easy-samba` development.
 
+### (2019-07-30 UTC) easy-samba 1.16.0 released (with new features)
+Stable version 1.16.0 of `easy-samba` has been released.
+
+This version brings new features to `easy-samba`.
+
+Click [here](https://github.com/adevur/docker-easy-samba/blob/master/docs/CHANGELOG.md#stable-feature-1160-2019-07-30-utc) for changelog.
+
 ### (2019-07-25 UTC) easy-samba 1.15.0 released (with new features)
 Stable version 1.15.0 of `easy-samba` has been released.